requirements.txt
faker
kafka-python==2.0.0
psycopg2==2.8.4
SQLAlchemy==1.3.13